revenge-server/revenge/requirements.txt
Django==1.9.6
pytz==2016.4
gunicorn==19.6.0
gevent==1.1.1
celery==3.1.24
django-celery==3.1.17
django-celery-email==1.1.5
pypinyin=0.17.0




